Extract performance workspace contract facade
